+++ src/messages/lib/MessageBalanceOf.js
@@ -1,0 +1,11 @@
+const AbstractMessage = require('./AbstractMessage')
+
+class MessageBalanceOf extends AbstractMessage {
+	constructor ({ balance }) {
+		super()
+		this.topic = 'balance_of'
+		this.balance = balance
+	}
+}
+
+module.exports = MessageBalanceOf